core/recommender.py
from utils.path_utils import CONFIG_PATH
import yaml

with open(CONFIG_PATH / "config.yaml") as f:
    THRESHOLDS = yaml.safe_load(f)

def generate_recommendations(risk_probability, student_input):

    observations = []
    actions = []

    if student_input["early_active_days"] == 0:
        observations.append("No engagement in early course period")
        actions.append("Start with introductory materials immendiately.")

    if student_input["early_click"] < THRESHOLDS["early_engagement"]["min_early_click"]:
        observations.append("Low early interaction with course content")
        actions.append("Watch intro videos and explore course structure")

    if student_input["first_activity_day"] > 0:
        observations.append(";ate course start compared to peers")
        actions.append("Follow a structured 7-day recovery plan")

    if risk_probability >= THRESHOLDS["risk_levels"]["high"]:
        actions.append("Schedule academic support or mentoring session.")

    return observations, actions
