requirements.txt
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0
joblib==1.5.3
streamlit==1.52.1
pyyaml==6.0.3